Update ProductActions.php
https://github.com/magento/magento2/issues/35783

This change was suggested to handle the name not being set on a product. A bad import could happen or the name could not be in the selected data or the name turned off from being required. Without this the catalog grid becomes unusable and left in a state that admin user cannot recover from.

Please make this change ASAP. Thanks.